Split activity_admin into it's own file
